test(reversi): reversi.spec.ts を責務ごとにファイル分割して並列実行する
1 ファイルにまとまっていた Board クラスのテストは Vitest のファイル並列実行の
恩恵を受けられなかった。メソッド単位でファイルを分割することで各ファイルが
独立したワーカーで並列実行され、テスト実行時間の短縮が見込める。
